feat: adicionando dialog para confirmação de exclusão ou cancelar preenchimento

+++ gerenciamento-chamados/src/app/chamados/pages/page-gerenciamento-chamados/page-gerenciamento-chamados.component.ts
@@ -6,6 +6,8 @@ import { ButtonComponent } from '../../../shared/components/button/button.compon
 import { BlocosChamadosComponent } from '../../components/blocos-chamados/blocos-chamados.component';
 import { CardChamadosComponent } from '../../components/card-chamados/card-chamados.component';
 import { ChamadosMockService } from '../../../services/chamados-mock.service';
+import { ConfirmDialogComponent } from '../../../shared/components/confirm-dialog/confirm-dialog.component';
+import { ConfirmDialogService } from '../../../shared/components/confirm-dialog/confirm-dialog.service';
 
 @Component({
   selector: 'app-page-gerenciamento-chamados',
@@ -17,12 +19,14 @@ import { ChamadosMockService } from '../../../services/chamados-mock.service';
     ButtonComponent,
     BlocosChamadosComponent,
     CardChamadosComponent,
+    ConfirmDialogComponent,
   ],
 })
 export class PageGerenciamentoChamadosComponent {
   constructor(
     private router: Router,
-    private chamadosService: ChamadosMockService
+    private chamadosService: ChamadosMockService,
+    private confirmDialogService: ConfirmDialogService
   ) {}
 
   chamadosNaoAtendidos = computed(() => this.chamadosService.byStatus('ATENDER'));
@@ -38,6 +42,14 @@ export class PageGerenciamentoChamadosComponent {
   }
 
   removerChamado(id: number): void {
-    this.chamadosService.remove(id);
+    this.confirmDialogService.confirm({
+      header: 'Excluir chamado?',
+      message: 'Tem certeza que deseja excluir este chamado?',
+      acceptLabel: 'Sim, excluir',
+      rejectLabel: 'Cancelar',
+      onAccept: () => {
+        this.chamadosService.remove(id);
+      }
+    });
   }
 }
